% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_size.R
\name{screen_size}
\alias{screen_size}
\title{Open a browser to get screen resolution}
\usage{
screen_size()
}
\description{
Open a browser to get screen resolution
}
\examples{
\dontrun{

screen_size()

}
}
